Code functionality
Using a light sensor with MicroPython to create an automatic nightlight on micro:bit. With using the light sensor to measure the surrounding intensity of light, I defined different light levels by MicroPython. The micro:bit output light through the LEDs built-in function. It vary with the surrounding intensity of light. When the surrounding intensity of light reaches the per-set levels, micro:bit will output the light which I defined, otherwise micro:bit will maintain a dimmer heart image as usual.